Add an `--overwrite` flag for enabling overwrite of existing chunk data
By default, the chunk exporter will fail if a directory already exists
for the chunk. This flag can be used to override this behavior and
replace the existing directory. For the production system, it is likely
that existing directories should not be overwritten, because this could
lead to duplicate data.