Revise chunk exporter to use new ApdbTableData interface
Instead of getting the column datatypes from the schema, the chunk
exporter will now read them directly from the `ApdbTableData` using the
`column_defs` method. A small module was added which provides the
mapping from the Felis `DataType` to arrow types, as well as a simple
function for creating an Arrow schema.