Implement database initialization for replica chunk tracking
This adds proper database initialization for replica chunk tracking
without including the instantiation of the schema. Only the
`PpdbReplicaChunk` and `metadata` tables are created.

Several utility modules were added to the `sql` package including
functionality currently define in `_ppdb_sql`. This was done so that
the classes in the `bigquery` module do not depend on `_ppdb_sql` any
longer, as it may be removed eventually.