Refactor methods used to query replica chunks in `PpdbReplicaChunkSql`
This reduces some code duplication in methods used to query for the
extended replica chunk information. The `Ppdb` interface was changed
slightly so that a `Sequence` of replica chunks is returned instead of
a `list`, allowing variadic typing so that sub-classes of
`PpdbReplicaChunk` may be used. It would work without this, but the
minor change to the interface fixes errors reported by mypy.