Make `PpdbBigQueryConfig` a direct sub-class of `PpdbConfig`
Instead of having `PpdbBigQueryConfig` sub-class `PpdbSqlConfig`, this
changes the class to be a child of `PpdbConfig`. The required
SQL configuration is included in its entirety under a `sql` key.

The new `PpdbSqlConfig` is also changed so that it has a
`apdb_schema_uri` attribute pointing to the APDB schema, with a default
resource path. The capability of loading a specific schema from a file
with multiple ones is removed, as this is not currently supported in
Felis.